Fixed amount to be an Integer, not a float

+++ android_studio_project/app/src/main/java/com/git_er_done/cmput301f22t06_team_project/models/DairyIngredient.java
@@ -4,7 +4,7 @@ import java.time.LocalDate;
 
 public class DairyIngredient extends Ingredient {
     public DairyIngredient(String name, String desc, LocalDate best_before, String location,
-                           String units, String category, float amount) {
+                           String units, String category, Integer amount) {
         super(name, desc, best_before, location, units, category, amount);
     }
 }
